services/ingestor/requirements.txt
boto3
pika
psycopg2-binary
python-dotenv